10455-story: new password in production should be randomized in add-user and add-judge
